requirements.txt
Django
djangorestframework
requests
#psycopg2
psycopg2-binary
